+++ src/sections/create-dataset/CreateDatasetFactory.tsx
@@ -1,18 +1,11 @@
-import { CreateDatasetFormPresenter } from './CreateDatasetForm'
-import { formValidationService } from '../../dataset/domain/useCases/createDataset'
-import { formSubmissionService } from '../../dataset/domain/useCases/createDataset'
+import { CreateDatasetForm } from './CreateDatasetForm'
 import { ReactElement } from 'react'
-export const CreateDatasetForm: React.FC = () => {
-  return (
-    <CreateDatasetFormPresenter
-      formValidationService={formValidationService}
-      formSubmissionService={formSubmissionService}
-    />
-  )
+export const CreateDatasetFormPresenter: React.FC = () => {
+  return <CreateDatasetForm />
 }
 
 export class CreateDatasetFactory {
   static create(): ReactElement {
-    return <CreateDatasetForm />
+    return <CreateDatasetFormPresenter />
   }
 }
