+++ tests/component/views/create-dataset/CreateDataset.spec.tsx
@@ -1,10 +1,9 @@
-// import DatasetCreate from '../../../../src/sections/create-dataset/CreateDatasetContext'
-import { CreateDatasetForm } from '../../../../src/sections/create-dataset/CreateDatasetFactory'
+import { CreateDatasetFormPresenter } from '../../../../src/sections/create-dataset/CreateDatasetFactory'
 import { mount } from 'cypress/react18'
 
 describe('Form component', () => {
   it('renders the Create Dataset page and its contents', () => {
-    mount(<CreateDatasetForm />)
+    cy.customMount(<CreateDatasetFormPresenter />)
     cy.findByText(/Create Dataset/i).should('exist')
 
     cy.findByLabelText(/Title/i).should('exist')
@@ -15,8 +14,8 @@ describe('Form component', () => {
   })
 
   it('can submit a valid form', () => {
-    cy.log('Submit form')
-    mount(<CreateDatasetForm />)
+    // cy.log('Submit form')
+    mount(<CreateDatasetFormPresenter />)
     cy.findByLabelText(/Title/i)
       .should('exist')
       .type('Test Dataset Title')
